Decouples InMemorySpanStore and tests from zipkin-server
This refactors InMemorySpanStore to not use Java 8 features. Now, it is
inside the core jar vs being stuck in the zipkin-server package.

The first consumer of this will be MockZipkinServer (#64)
