+++ zipkin/src/test/java/zipkin/DependenciesTest.java
